+++ sekoia_automation/exceptions.py
@@ -9,7 +9,9 @@ class AutomationSDKError(Exception):
 
 
 class SendEventError(AutomationSDKError):
-    pass
+    def __init__(self, message: str, status_code: int = 0):
+        super().__init__(message)
+        self.status_code = status_code
 
 
 class CommandNotFoundError(AutomationSDKError):